feature: Moving logs utils and batch processor to utils dir and typo fixes in docs (#1191)
